week2_결과물/requirements.txt
langchain>=1.0.0
langchain-core>=0.3.0
langchain-groq>=0.3.0
langchain-ollama>=0.2.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
pandas>=2.0.0
python-dotenv>=1.0.0
